refactor : adapter in/out 분리
