+++ user-service/src/main/java/com/gaebal_easy/client/user/presentation/adapter/in/DeliveryUserEventListener.java
@@ -1,6 +1,5 @@
-package com.gaebal_easy.client.user.presentation;
+package com.gaebal_easy.client.user.presentation.adapter.in;
 
-import com.gaebal_easy.client.user.application.dto.HubManagerInfoMessage;
 import com.gaebal_easy.client.user.application.service.DeliveryUserEventService;
 import gaebal_easy.common.global.message.DeliveryUserDeleteMessage;
 import gaebal_easy.common.global.message.DeliveryUserInfoMessage;
